requirements.txt
black
py-solc-x
python-dotenv
eth-brownie
pytest